+++ src/components/TestList.js
@@ -1,0 +1,16 @@
+import React, { Component } from 'react';
+
+class TestList extends Component {
+    render() {
+        return (
+            <div>
+                <h4>simple List</h4>
+                <ul>
+                    <li>list element</li>
+                </ul>
+            </div>
+        );
+    }
+};
+
+export default TestList;