+++ src/reducers/index.js
@@ -1,0 +1,7 @@
+import { combineReducers } from 'redux';
+import authReducer from './simpleAuthButton';
+
+export default combineReducers({
+  //- put reducer key: functioncalls here
+  auth: authReducer
+});