+++ CHANGELOG.md
@@ -1,9 +1,13 @@
 # Changelog
 All notable changes to this project will be documented in this file.
 
+## [0.0.3] - 2019-04-25
+
+- Create simple spa with nav and 100% test coverage [v0.0.3-ReleaseNotes.md](docs/releases/v0.0.3-ReleaseNotes.md)
+
 ## [0.0.2] - 2019-04-24
 
-- Create basic react starter app
+- Create basic react starter app see [v0.0.2-ReleaseNotes.md](docs/releases/v0.0.2-ReleaseNotes.md)
 
 ## [0.0.1] - 2019-04-24